Refactor and formatting
