+++ src/main/java/org/alliancegenome/curation_api/controllers/document/AlleleDocumentController.java
@@ -1,16 +1,22 @@
 package org.alliancegenome.curation_api.controllers.document;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 import org.alliancegenome.curation_api.interfaces.document.AlleleDocumentInterface;
 import org.alliancegenome.curation_api.model.document.builders.AlleleSummaryDocumentBuilder;
+import org.alliancegenome.curation_api.model.document.builders.TransgenicAlleleDocumentBuilder;
 import org.alliancegenome.curation_api.model.document.es.AlleleSummaryDTO;
 import org.alliancegenome.curation_api.model.document.es.AlleleSummaryDocument;
-import org.alliancegenome.curation_api.response.ObjectListResponse;
+import org.alliancegenome.curation_api.model.document.es.TransgenicAlleleDTO;
+import org.alliancegenome.curation_api.model.entities.associations.AlleleConstructAssociation;
+import org.alliancegenome.curation_api.model.input.Pagination;
 import org.alliancegenome.curation_api.response.SearchResponse;
 import org.alliancegenome.curation_api.services.AlleleService;
 import org.alliancegenome.curation_api.services.ResourceDescriptorPageService;
+import org.alliancegenome.curation_api.services.associations.AlleleConstructAssociationService;
+
 import jakarta.inject.Inject;
 import lombok.extern.jbosslog.JBossLog;
 
@@ -19,7 +25,10 @@ public class AlleleDocumentController implements AlleleDocumentInterface {
 
 	@Inject
 	AlleleService alleleService;
-
+	
+	@Inject
+	AlleleConstructAssociationService acService;
+	
 	@Inject
 	ResourceDescriptorPageService resourceDescriptorPageService;
 
@@ -27,7 +36,7 @@ public class AlleleDocumentController implements AlleleDocumentInterface {
 	public SearchResponse<Long> getAllIds() {
 		List<Long> ids = alleleService.getAllAlleleSummaryIds();
 		SearchResponse<Long> response = new SearchResponse<>(ids);
-		response.setTotalResults((long) ids.size());
+		response.setTotalResults(ids.size());
 		return response;
 	}
 
@@ -45,6 +54,28 @@ public class AlleleDocumentController implements AlleleDocumentInterface {
 
 		SearchResponse<AlleleSummaryDocument> ret = new SearchResponse<>(list);
 		ret.setTotalResults(resp.getTotalResults());
+		return ret;
+	}
+
+	@Override
+	public SearchResponse<TransgenicAlleleDTO> findDocuments(Integer page, Integer limit, HashMap<String, Object> params) {
+		if (params == null) {
+			params = new HashMap<>();
+		}
+		params.put("alleleAssociationSubject.obsolete", false);
+		params.put("alleleAssociationSubject.internal", false);
+
+		SearchResponse<AlleleConstructAssociation> response = acService.findByParams(new Pagination(page, limit), params);
+		List<TransgenicAlleleDTO> list = new ArrayList<>();
+		if (response.getResults() != null) {
+			TransgenicAlleleDocumentBuilder builder = new TransgenicAlleleDocumentBuilder();
+			for (AlleleConstructAssociation association : response.getResults()) {
+				TransgenicAlleleDTO doc = builder.buildTransgenicAlleleDocument(association);
+				list.add(doc);
+			}
+		}
+		SearchResponse<TransgenicAlleleDTO> ret = new SearchResponse<>(list);
+		ret.setTotalResults(response.getTotalResults());
 		return ret;
 	}
